[REFACTORING] Moved some classes around + added a footer to the CLI about config files
