+++ src/main/java/life/qbic/model/UnitConverter/UnitDisplay.java
@@ -1,4 +1,4 @@
-package life.qbic.UnitConverter;
+package life.qbic.model.UnitConverter;
 
 public interface UnitDisplay {
